refactor: refactor package location
