+++ src/main/kotlin/com/oneplace/dgapiserver/domain/account/application/AccountService.kt
@@ -1,4 +1,4 @@
-package com.oneplace.dgapiserver.domain.auth
+package com.oneplace.dgapiserver.domain.account.application
 
 import org.springframework.http.ResponseEntity
