+++ src/main/kotlin/com/oneplace/dgapiserver/domain/account/domain/repository/AccountRepository.kt
@@ -1,5 +1,6 @@
-package com.oneplace.dgapiserver.domain.auth
+package com.oneplace.dgapiserver.domain.account.domain.repository
 
+import com.oneplace.dgapiserver.domain.account.domain.Account
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface AccountRepository : JpaRepository<Account, Long> {
